Rename backend metadata headers to "meta.h"
