+++ examples/monolithic_build/config_1024.h
@@ -88,7 +88,7 @@
  *****************************************************************************/
 #if defined(MLKEM_USE_NATIVE_BACKEND_ARITH) && \
     !defined(MLKEM_NATIVE_ARITH_BACKEND_FILE)
-#define MLKEM_NATIVE_ARITH_BACKEND_FILE "native/default.h"
+#define MLKEM_NATIVE_ARITH_BACKEND_FILE "native/meta.h"
 #endif
 
 /******************************************************************************
@@ -128,7 +128,7 @@
  *****************************************************************************/
 #if defined(MLKEM_USE_NATIVE_BACKEND_FIPS202) && \
     !defined(MLKEM_NATIVE_FIPS202_BACKEND_FILE)
-#define MLKEM_NATIVE_FIPS202_BACKEND_FILE "fips202/native/default.h"
+#define MLKEM_NATIVE_FIPS202_BACKEND_FILE "fips202/native/meta.h"
 #endif
 
 #endif /* MLkEM_NATIVE_CONFIG_H */
