+++ mlkem/fips202/native/x86_64/meta.h
@@ -3,8 +3,8 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-#ifndef MLKEM_NATIVE_FIPS202_NATIVE_X86_64_XKCP_H
-#define MLKEM_NATIVE_FIPS202_NATIVE_X86_64_XKCP_H
+#ifndef MLKEM_NATIVE_FIPS202_NATIVE_X86_64_META_H
+#define MLKEM_NATIVE_FIPS202_NATIVE_X86_64_META_H
 /* Default FIPS202 assembly profile for AArch64 systems */
 
 #ifdef MLKEM_NATIVE_FIPS202_PROFILE_H
@@ -26,4 +26,4 @@
 
 #endif /* MLKEM_NATIVE_FIPS202_PROFILE_H */
 
-#endif /* MLKEM_NATIVE_FIPS202_NATIVE_X86_64_XKCP_H */
+#endif /* MLKEM_NATIVE_FIPS202_NATIVE_X86_64_META_H */
