+++ mlkem/native/meta.h
@@ -2,8 +2,8 @@
  * Copyright (c) 2024-2025 The mlkem-native project authors
  * SPDX-License-Identifier: Apache-2.0
  */
-#ifndef MLKEM_NATIVE_NATIVE_DEFAULT_H
-#define MLKEM_NATIVE_NATIVE_DEFAULT_H
+#ifndef MLKEM_NATIVE_NATIVE_META_H
+#define MLKEM_NATIVE_NATIVE_META_H
 
 /*
  * Default arithmetic backend
@@ -11,11 +11,11 @@
 #include "../sys.h"
 
 #ifdef SYS_AARCH64
-#include "aarch64/opt.h"
+#include "aarch64/meta.h"
 #endif /* SYS_AARCH64 */
 
 #ifdef SYS_X86_64_AVX2
-#include "x86_64/default.h"
+#include "x86_64/meta.h"
 #endif /* SYS_X86_64 */
 
-#endif /* MLKEM_NATIVE_NATIVE_DEFAULT_H */
+#endif /* MLKEM_NATIVE_NATIVE_META_H */
